refactor: countPublicReportsByBillId を @mirai-gikai/shared に集約
muraikenta レビュー(フォローアップ)対応。public-read-repository.ts に追加した
重複コピーを削除し、packages/shared/src/report-publication/count-public-reports.ts に
共通化する。web の interview-report-repository.ts は re-export に切り替え、既存6箇所の
呼び出しは無変更。ブラウザ安全な auto-publish.ts とは別ファイル/別 export サブパスに
分け、server-only + createAdminClient をクライアント境界に持ち込まない。
クエリは完全同一で挙動不変。
